LayoutTests/svg/dynamic-updates/SVGEllipseElement-svgdom-requiredFeatures.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="resources/SVGTestCase.js"></script>
<script src="../../resources/js-test-pre.js"></script>
<script src="../../fast/repaint/resources/repaint.js"></script>
</head>
<body onload="runRepaintTest()">
<h1>SVG 1.1 dynamic update tests</h1>
<p id="description"></p>
<div id="console"></div>
<script>
// [Name] SVGEllipseElement-svgdom-requiredFeatures.js
// [Expected rendering result] a series of PASS messages

createSVGTestCase();

var ellipseElement = createSVGElement("ellipse");
ellipseElement.setAttribute("rx", "200");
ellipseElement.setAttribute("ry", "200");

rootSVGElement.appendChild(ellipseElement);

function repaintTest() {
    debug("Check that SVGEllipseElement is initially displayed");
    shouldBeEqualToString("document.defaultView.getComputedStyle(ellipseElement, null).display", "inline");
    debug("Check that setting requiredFeatures to something invalid makes it not render");
    ellipseElement.requiredFeatures.appendItem("http://www.w3.org/TR/SVG11/feature#BogusFeature");
    shouldBeEqualToString("document.defaultView.getComputedStyle(ellipseElement, null).display", "");
    debug("Check that setting requiredFeatures to something valid makes it render again");
    ellipseElement.requiredFeatures.replaceItem("http://www.w3.org/TR/SVG11/feature#Shape", 0);
    shouldBeEqualToString("document.defaultView.getComputedStyle(ellipseElement, null).display", "inline");
    debug("Check that adding something valid to requiredFeatures keeps rendering the element");
    ellipseElement.requiredFeatures.appendItem("http://www.w3.org/TR/SVG11/feature#Gradient");
    shouldBeEqualToString("document.defaultView.getComputedStyle(ellipseElement, null).display", "inline");
    debug("Check that adding something invalid to requiredFeatures makes it not render");
    ellipseElement.requiredFeatures.appendItem("http://www.w3.org/TR/SVG11/feature#BogusFeature");
    shouldBeEqualToString("document.defaultView.getComputedStyle(ellipseElement, null).display", "");

    completeTest();
}

var successfullyParsed = true;
</script>
</body>
</html>
